code motion: factor our execution::types module
