+++ rust/kcl-lib/src/docs/gen_std_tests.rs
@@ -23,7 +23,9 @@ use crate::{
 const TYPES_DIR: &str = "../../docs/kcl/types";
 const LANG_TOPICS: [&str; 5] = ["Types", "Modules", "Settings", "Known Issues", "Constants"];
 // These types are declared in std.
-const DECLARED_TYPES: [&str; 9] = ["number", "string", "tag", "bool", "Sketch", "Solid", "Plane", "Helix", "Face"];
+const DECLARED_TYPES: [&str; 9] = [
+    "number", "string", "tag", "bool", "Sketch", "Solid", "Plane", "Helix", "Face",
+];
 
 fn init_handlebars() -> Result<handlebars::Handlebars<'static>> {
     let mut hbs = handlebars::Handlebars::new();
